+++ lib/strong_migrations/safe_methods.rb
@@ -69,9 +69,14 @@ module StrongMigrations
       end
     end
 
-    def safe_change_column_null(add_code, validate_code, change_args, remove_code)
+    def safe_change_column_null(add_code, validate_code, change_args, remove_code, default)
       @migration.reversible do |dir|
         dir.up do
+          # TODO raise error if not supported in 0.9.0
+          unless default.nil?
+            warn "[strong_migrations] default value not supported yet"
+          end
+
           @migration.safety_assured do
             @migration.execute(add_code)
             disable_transaction